withdrawns migration

+++ migrations/sql/20250630195349_create_orders.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS orders
 (
     number   text PRIMARY KEY,
     status   text NOT NULL,
-    accrual  bigint DEFAULT 0,
+    accrual  DECIMAL(10,5) DEFAULT 0,
     uploaded_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
     user_id  bigint NOT NULL
 );
